requirements.txt
arrow==0.15.5
asgiref==3.2.3
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.1
blessed==1.17.1
bootstrap-admin==0.4.3
cffi==1.13.2
confusable-homoglyphs==3.2.0
cryptography==2.8
decorator==4.4.1
Django==2.2
django-appconf==1.0.3
django-authority==0.14
django-autoslug==1.9.6
django-bootstrap4==1.1.1
django-ckeditor==5.8.0
django-countries==5.5
django-cryptography==0.4
django-extensions==2.2.6
django-grappelli==2.13.3
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-picklefield==2.1.1
django-q==1.1.0
django-qr-code==1.1.0
django-registration-redux==2.6
django-simple-history==2.8.0
easy-thumbnails==2.7
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.2
pexpect==4.7.0
phonenumbers==8.11.1
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
python-crontab==1.9.3
python-dateutil==2.8.1
pytz==2019.3
qrcode==6.1
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
